pyproject.toml
[project]
name = "gis-dashboard-dt-parent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "truststore>=0.10.4",
]
